Fix minor bugs

+++ WordWeaver/Data/Entity/UserRole.cs
@@ -7,7 +7,7 @@ public partial class UserRole
 {
     public long UserRoleId { get; set; }
 
-    public long? RoleId { get; set; }
+    public int? RoleId { get; set; }
 
     public long? UserId { get; set; }
